Fix warnings, encodings

+++ XYmap.h
@@ -48,14 +48,14 @@ CRGB leds[ NUM_LEDS ];
 // This code, plus the supporting 80-byte table is much smaller 
 // and much faster than trying to calculate the pixel ID with code.
 #define LAST_VISIBLE_LED 67
-uint8_t XY( uint8_t x, uint8_t y)
+uint16_t XY( uint16_t x, uint16_t y)
 {
   // any out of bounds address maps to the first hidden pixel
   if( (x >= kMatrixWidth) || (y >= kMatrixHeight) ) {
     return (LAST_VISIBLE_LED + 1);
   }
 
-  const uint8_t ShadesTable[] = {
+  const uint16_t ShadesTable[] = {
      68,  0,  1,  2,  3,  4,  5,  6,  7,  8,  9, 10, 11, 12, 13, 69,
      29, 28, 27, 26, 25, 24, 23, 22, 21, 20, 19, 18, 17, 16, 15, 14,
      30, 31, 32, 33, 34, 35, 36, 70, 71, 37, 38, 39, 40, 41, 42, 43,
@@ -63,9 +63,7 @@ uint8_t XY( uint8_t x, uint8_t y)
      74, 58, 59, 60, 61, 62, 75, 76, 77, 78, 63, 64, 65, 66, 67, 79
   };
 
-  uint8_t i = (y * kMatrixWidth) + x;
-  uint8_t j = ShadesTable[i];
+  uint16_t i = (y * kMatrixWidth) + x;
+  uint16_t j = ShadesTable[i];
   return j;
 }
-
-
